+++ test/test_handsome.ml
@@ -4,5 +4,6 @@ let () =
     ([ Test_doc.suite; Test_render.suite; Test_diff.suite ]
      @ Test_laws.suites
      @ Test_width.suites
-     @ [ Test_framed.suite ])
+     @ [ Test_framed.suite ]
+     @ Test_fit.suites)
 ;;